services/vectorstore_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sentence-transformers==2.6.1
faiss-cpu==1.8.0
pydantic==2.6.4
numpy==1.26.4
